applied changes, improved congfig file + encapsulated logic in genesis.rs
