--- genesis.toml
@@ -1,19 +1,0 @@
-version = 1
-timestamp = 1672531200
-
-[[accounts]]
-type = "BasicWallet"
-mode = "RegularAccountImmutableCode"
-account_seed = "a123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef"
-auth_seed = "b123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef"
-auth_scheme = "RpoFalcon512"
-
-[[accounts]]
-type = "BasicFungibleFaucet"
-mode = "FungibleFaucet"
-account_seed = "c123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef"
-auth_scheme = "RpoFalcon512"
-auth_seed = "d123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef"
-token_symbol = "POL"
-decimals = 12
-max_supply = 1000000
